Co-locate every test with the module it tests
AGENTS.md rule 11 requires each module `foo.py` to have its tests in a
sibling `foo_test.py` -- no `tests/` directory, no `test_*.py` prefix.
Two categories of violation remained.

Prefix violation: `_csr/test_util.py` matched pytest's default
`test_*.py` collection glob. Renamed to `_csr/_test_util.py` alongside
its test, five importers, and the coverage omit glob.

Orphan tests (a `foo_test.py` with no sibling `foo.py`): eleven files.
Eight merge straight into the module they actually exercise; two
policy-oriented suites (`indptr_dtype_test.py`, `cuda_int64_indptr_test.py`)
split by target module, since they were organised by feature rather than
by module. Two of those splits land as the first tests for `diag_add.py`
and `hybrid_config.py`.

`deprecation_test.py` had no legal sibling -- it tests the `__getattr__`
shim in `__init__.py`. The shim moves to a new `brainevent/_deprecation.py`
(the only shipping-code change here). The rename table now stores target
*name strings* resolved against a caller-supplied namespace rather than
live objects, which avoids a circular import and makes `resolve` unit
testable directly. Public behaviour is unchanged.

Three merge destinations carried a module-level
`pytestmark = pytest.mark.slow`; appending fast tests there would have
silently dropped 152 fast tests out of the default lane while still
collecting under `-m ''`. Those blanket marks are now per-item
decorators. Two analogous module-scope hazards in the `_op` merge (a
Windows `allow_module_level` skip, a never-restoring autouse x64 fixture)
are narrowed the same way.

Collection: 3636 -> 3648 full, 1775 -> 1787 fast. The +12 is 10 new
`_deprecation` unit tests plus 2 from splitting one cross-module test in
three. `pytest -m ''` passes 3426 with 0 failures.

Spec: docs/specs/test-colocation.md
