+++ brainevent/_csr/_test_util.py
@@ -19,6 +19,7 @@ import brainstate
 import jax
 import jax.numpy as jnp
 import numpy as np
+import pytest
 
 
 def get_csr(n_pre, n_post, prob, replace=True):
@@ -251,3 +252,131 @@ def _csr_matrix_impl(xs, w, indices, indptr, shape, n_conn):
         return out.at[i].set(jnp.sum(xs[ids] * ws, axis=0))
 
     return jax.lax.fori_loop(0, shape[0], body_fn, out)
+
+
+def small_csr():
+    """Return a tiny 2x3 CSR with int32 structure, for dtype/structure tests.
+
+    Returns
+    -------
+    brainevent.CSR
+        The matrix ``[[1, 0, 2], [0, 3, 0]]`` with float32 data and int32
+        ``indices``/``indptr``.
+
+    Examples
+    --------
+    .. code-block:: python
+
+        >>> from brainevent._csr._test_util import small_csr
+        >>> small_csr().shape
+        (2, 3)
+    """
+    import brainevent
+
+    data = jnp.array([1.0, 2.0, 3.0], dtype=jnp.float32)
+    indices = jnp.array([0, 2, 1], dtype=jnp.int32)
+    indptr = jnp.array([0, 2, 3], dtype=jnp.int32)
+    return brainevent.CSR((data, indices, indptr), shape=(2, 3))
+
+
+# ---------------------------------------------------------------------------
+# Helpers for the CUDA int64-indptr policy tests, shared by the per-module
+# test files (binary, binary_indexed, float, slice, dt2t, plasticity_binary).
+# ---------------------------------------------------------------------------
+
+#: Skip when no GPU backend is active. Distinct from
+#: ``brainevent._test_util.requires_gpu``: that one probes ``jax.devices('gpu')``,
+#: this one requires the *default* backend to be the GPU, because these tests
+#: compare a ``cuda_raw`` result against a ``jax_raw`` reference.
+requires_gpu_backend = pytest.mark.skipif(
+    jax.default_backend() != 'gpu',
+    reason='CUDA int64 indptr tests require a GPU backend',
+)
+
+
+def strip_hybrid_suffix(name):
+    """Remove the config-dependent hybrid module suffix (e.g. ``_b256_s2048_t128_n4096``).
+
+    Parameters
+    ----------
+    name : str
+        A compiled-module or FFI-target name, possibly carrying the suffix.
+
+    Returns
+    -------
+    str
+        ``name`` with the active hybrid-config suffix removed.
+
+    Notes
+    -----
+    The hybrid ``.so`` module names embed the active
+    :func:`~brainevent._csr.hybrid_config.get_hybrid_config` so different configs
+    compile to distinct modules. These tests only care about the logical module
+    identity, so the suffix is normalised away to stay config-agnostic.
+    """
+    from brainevent._csr.hybrid_config import get_hybrid_config, module_suffix_for_config
+
+    suffix = module_suffix_for_config(get_hybrid_config())
+    return name.replace(suffix, '') if suffix else name
+
+
+def int64_structure(indptr_dtype):
+    """Return a tiny ``(weights, indices, indptr)`` triple with the given indptr dtype.
+
+    Parameters
+    ----------
+    indptr_dtype : dtype
+        Dtype for the returned ``indptr`` array.
+
+    Returns
+    -------
+    tuple
+        ``(weights, indices, indptr)`` describing a 2x3 CSR with 4 stored entries.
+    """
+    weights = jnp.array([1.0, 2.0, 3.0, 4.0], dtype=jnp.float32)
+    indices = jnp.array([0, 2, 1, 2], dtype=jnp.int32)
+    indptr = jnp.array([0, 2, 4], dtype=indptr_dtype)
+    return weights, indices, indptr
+
+
+def shape_of(dtype, shape=(2,)):
+    """Return a :class:`jax.ShapeDtypeStruct` for ``shape``/``dtype``."""
+    return jax.ShapeDtypeStruct(shape, dtype)
+
+
+def cuda_kwargs(indices_dtype=jnp.int64, indptr_dtype=jnp.int64, nse=2):
+    """Return the common kwargs a CSR CUDA kernel generator expects."""
+    return {
+        'outs': [shape_of(jnp.float32)],
+        'shape': (1, 2),
+        'indices_info': shape_of(indices_dtype, (nse,)),
+        'indptr_info': shape_of(indptr_dtype, (2,)),
+    }
+
+
+def recording_ffi_call(calls):
+    """Return a ``jax.ffi.ffi_call`` stand-in that records invocations into ``calls``.
+
+    Parameters
+    ----------
+    calls : list
+        Mutated in place; each entry is
+        ``(name, out_info, ffi_kwargs, args, kwargs)``.
+
+    Returns
+    -------
+    callable
+        A drop-in replacement for ``jax.ffi.ffi_call`` returning zeros of the
+        requested output shape, so a kernel can be exercised without a GPU.
+    """
+
+    def ffi_call(name, out_info, **ffi_kwargs):
+        def call(*args, **kwargs):
+            calls.append((name, out_info, ffi_kwargs, args, kwargs))
+            if isinstance(out_info, (tuple, list)):
+                return tuple(jnp.zeros(info.shape, info.dtype) for info in out_info)
+            return jnp.zeros(out_info.shape, out_info.dtype)
+
+        return call
+
+    return ffi_call
