+++ brainevent/_csr/binary_indexed_test.py
@@ -30,7 +30,13 @@ from brainevent._csr.binary_indexed import (
     binary_csrmm_indexed,
     binary_csrmm_indexed_p_call,
 )
+import brainevent._csr.binary_indexed as binary_indexed_mod
 from brainevent._csr.main import _binary_workspace, _make_binary_task_workspace
+from brainevent._csr._test_util import (
+    cuda_kwargs, int64_structure, recording_ffi_call, requires_gpu_backend,
+    shape_of, strip_hybrid_suffix,
+)
+from brainevent._test_util import jax_x64_enabled
 from brainevent import BinaryArray, CSC, CSR, FixedNumPerPre
 
 
@@ -860,3 +866,261 @@ def test_indexed_mm_is_exported():
     assert "binary_csrmm_indexed" in brainevent.__all__
     assert "binary_csrmm_indexed_p" in brainevent.__all__
     assert binary_csrmm_indexed is brainevent.binary_csrmm_indexed
+
+
+# ---------------------------------------------------------------------------
+# int64 ``indptr`` policy on the CUDA path.
+#
+# ``indices`` stay int32 (the CUDA ABI is int32-only for coordinates) while
+# ``indptr`` may widen to int64. The generator tests run without a real GPU by
+# stubbing ``load_cuda_file``/``ffi_call``; the ``accepts`` tests need one.
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.parametrize(
+    'factory,args,kwargs',
+    [
+        (
+            binary_indexed_mod._binary_csrmv_indexed_cuda_kernel,
+            (shape_of(jnp.float32), shape_of(jnp.bool_), False),
+            {'outs': [shape_of(jnp.float32)], 'perm_info': shape_of(jnp.int32)},
+        ),
+        (
+            binary_indexed_mod._binary_csrmm_indexed_cuda_kernel,
+            (shape_of(jnp.float32), shape_of(jnp.bool_, (2, 1)), False),
+            {'outs': [shape_of(jnp.float32, (1, 1))], 'perm_info': shape_of(jnp.int32)},
+        ),
+    ],
+)
+def test_cuda_kernel_generators_reject_int64_indices_before_loading_cuda(factory, args, kwargs):
+    call_kwargs = cuda_kwargs()
+    call_kwargs.update(kwargs)
+
+    with pytest.raises(TypeError, match="indices with dtype int32"):
+        factory(*args, **call_kwargs)
+
+
+def test_binary_indexed_cuda_generators_accept_int64_indptr_without_real_cuda(monkeypatch):
+    ffi_calls = []
+    load_calls = []
+
+    monkeypatch.setattr(binary_indexed_mod, "load_cuda_file", lambda path, name, **kwargs: load_calls.append((path, name, kwargs)))
+    monkeypatch.setattr(binary_indexed_mod.jax.ffi, "ffi_call", recording_ffi_call(ffi_calls))
+
+    with jax_x64_enabled():
+        indices = jnp.array([0, 1], dtype=jnp.int32)
+        indptr = jnp.array([0, 2], dtype=jnp.int64)
+        perm = jnp.array([1, 0], dtype=jnp.int32)
+        workspace = _make_binary_task_workspace(indptr)
+        task_kwargs = {
+            'task_begin_info': shape_of(workspace.task_begin.dtype, workspace.task_begin.shape),
+            'task_end_info': shape_of(workspace.task_end.dtype, workspace.task_end.shape),
+            'status_info': shape_of(workspace.status.dtype, workspace.status.shape),
+            'task_capacity': workspace.task_capacity,
+        }
+        mv_task_outs = (
+            shape_of(jnp.float32),
+            task_kwargs['task_begin_info'],
+            task_kwargs['task_end_info'],
+            task_kwargs['status_info'],
+        )
+        mm_nt_task_outs = (
+            shape_of(jnp.float32, (1, 1)),
+            task_kwargs['task_begin_info'],
+            task_kwargs['task_end_info'],
+            task_kwargs['status_info'],
+        )
+        mm_t_task_outs = (
+            shape_of(jnp.float32, (2, 1)),
+            task_kwargs['task_begin_info'],
+            task_kwargs['task_end_info'],
+            task_kwargs['status_info'],
+        )
+
+        mv_kernel = binary_indexed_mod._binary_csrmv_indexed_cuda_kernel(
+            shape_of(jnp.float32, (2,)),
+            shape_of(jnp.bool_, (2,)),
+            False,
+            **{
+                **cuda_kwargs(indices_dtype=jnp.int32, indptr_dtype=jnp.int64),
+                'perm_info': shape_of(jnp.int32, (2,)),
+                **task_kwargs,
+            },
+        )
+        mv_kernel(
+            jnp.array([2.0, 3.0], dtype=jnp.float32),
+            indices,
+            indptr,
+            perm,
+            jnp.array([True, False]),
+            workspace.task_begin,
+            workspace.task_end,
+            workspace.status,
+        )
+
+        mv_t_kernel = binary_indexed_mod._binary_csrmv_indexed_cuda_kernel(
+            shape_of(jnp.float32, (2,)),
+            shape_of(jnp.bool_, (1,)),
+            True,
+            **{
+                **cuda_kwargs(indices_dtype=jnp.int32, indptr_dtype=jnp.int64),
+                'perm_info': shape_of(jnp.int64, (2,)),
+                'outs': mv_task_outs,
+                **task_kwargs,
+            },
+        )
+        mv_t_kernel(
+            jnp.array([2.0, 3.0], dtype=jnp.float32),
+            indices,
+            indptr,
+            perm.astype(jnp.int64),
+            jnp.array([True]),
+            workspace.task_begin,
+            workspace.task_end,
+            workspace.status,
+        )
+
+        mv_t_homo_kernel = binary_indexed_mod._binary_csrmv_indexed_cuda_kernel(
+            shape_of(jnp.float32, (1,)),
+            shape_of(jnp.bool_, (1,)),
+            True,
+            **{
+                **cuda_kwargs(indices_dtype=jnp.int32, indptr_dtype=jnp.int64),
+                'perm_info': shape_of(jnp.int64, (2,)),
+                'outs': mv_task_outs,
+                **task_kwargs,
+            },
+        )
+        mv_t_homo_kernel(
+            jnp.array([2.0], dtype=jnp.float32),
+            indices,
+            indptr,
+            perm.astype(jnp.int64),
+            jnp.array([True]),
+            workspace.task_begin,
+            workspace.task_end,
+            workspace.status,
+        )
+
+        mm_nt_kernel = binary_indexed_mod._binary_csrmm_indexed_cuda_kernel(
+            shape_of(jnp.float32, (2,)),
+            shape_of(jnp.bool_, (2, 1)),
+            False,
+            **{
+                **cuda_kwargs(indices_dtype=jnp.int32, indptr_dtype=jnp.int64),
+                'outs': mm_nt_task_outs,
+                'perm_info': shape_of(jnp.int32, (2,)),
+                **task_kwargs,
+            },
+        )
+        mm_nt_kernel(
+            jnp.array([2.0, 3.0], dtype=jnp.float32),
+            indices,
+            indptr,
+            perm,
+            jnp.array([[True], [False]]),
+            workspace.task_begin,
+            workspace.task_end,
+            workspace.status,
+        )
+
+        mm_t_kernel = binary_indexed_mod._binary_csrmm_indexed_cuda_kernel(
+            shape_of(jnp.float32, (2,)),
+            shape_of(jnp.bool_, (1, 1)),
+            True,
+            **{
+                **cuda_kwargs(indices_dtype=jnp.int32, indptr_dtype=jnp.int64),
+                'outs': mm_t_task_outs,
+                'perm_info': shape_of(jnp.int64, (2,)),
+                **task_kwargs,
+            },
+        )
+        mm_t_kernel(
+            jnp.array([2.0, 3.0], dtype=jnp.float32),
+            indices,
+            indptr,
+            perm.astype(jnp.int64),
+            jnp.array([[True]]),
+            workspace.task_begin,
+            workspace.task_end,
+            workspace.status,
+        )
+
+        mm_t_homo_kernel = binary_indexed_mod._binary_csrmm_indexed_cuda_kernel(
+            shape_of(jnp.float32, (1,)),
+            shape_of(jnp.bool_, (1, 1)),
+            True,
+            **{
+                **cuda_kwargs(indices_dtype=jnp.int32, indptr_dtype=jnp.int64),
+                'outs': mm_t_task_outs,
+                'perm_info': shape_of(jnp.int64, (2,)),
+                **task_kwargs,
+            },
+        )
+        mm_t_homo_kernel(
+            jnp.array([2.0], dtype=jnp.float32),
+            indices,
+            indptr,
+            perm.astype(jnp.int64),
+            jnp.array([[True]]),
+            workspace.task_begin,
+            workspace.task_end,
+            workspace.status,
+        )
+
+    assert [strip_hybrid_suffix(name) for _, name, _ in load_calls] == [
+        'csr_binary_indexed_csrmv',
+        'csr_binary_indexed_csrmv_hybrid',
+        'csr_binary_csrmv_hybrid',
+        'csr_binary_indexed_csrmm',
+        'csr_binary_indexed_csrmm_hybrid',
+        'csr_binary_csrmm_hybrid',
+    ]
+    assert [strip_hybrid_suffix(call[0]) for call in ffi_calls] == [
+        'csr_binary_indexed_csrmv.binary_csrmv_nt_auto_perm_hetero_f32_bool',
+        'csr_binary_indexed_csrmv_hybrid.binary_indexed_csrmv_wat_hybrid_hetero_f32_bool',
+        'csr_binary_csrmv_hybrid.binary_csrmv_wat_hybrid_homo_f32_bool',
+        'csr_binary_indexed_csrmm.binary_csrmm_nt_auto_perm_hetero_f32_bool',
+        'csr_binary_indexed_csrmm_hybrid.binary_indexed_csrmm_sraw_hybrid_hetero_f32_bool',
+        'csr_binary_csrmm_hybrid.binary_csrmm_sraw_hybrid_homo_f32_bool',
+    ]
+
+
+@requires_gpu_backend
+@pytest.mark.parametrize('transpose', [False, True])
+def test_binary_indexed_cuda_accepts_int64_indptr(transpose):
+    weights, indices, indptr32 = int64_structure(jnp.int32)
+    indptr64 = indptr32.astype(jnp.int64)
+    perm = jnp.array([2, 0, 3, 1], dtype=jnp.int32)
+    vector = jnp.array([True, False], dtype=jnp.bool_) if transpose else jnp.array([True, False, True])
+    workspace64 = _make_binary_task_workspace(indptr64)
+    workspace32 = _make_binary_task_workspace(indptr32)
+
+    got = binary_csrmv_indexed(weights, indices, indptr64, perm, vector, shape=(2, 3),
+                               transpose=transpose, backend='cuda_raw', workspace=workspace64)
+    expected = binary_csrmv_indexed(weights, indices, indptr32, perm, vector, shape=(2, 3),
+                                    transpose=transpose, backend='jax_raw', workspace=workspace32)
+
+    assert jnp.allclose(got, expected, rtol=1e-5, atol=1e-5)
+
+
+@requires_gpu_backend
+@pytest.mark.parametrize('transpose', [False, True])
+def test_binary_indexed_csrmm_cuda_accepts_int64_indptr(transpose):
+    weights, indices, indptr32 = int64_structure(jnp.int32)
+    indptr64 = indptr32.astype(jnp.int64)
+    perm = jnp.array([2, 0, 3, 1], dtype=jnp.int32)
+    matrix = (
+        jnp.array([[True, False], [False, True]], dtype=jnp.bool_)
+        if transpose else
+        jnp.array([[True, False], [False, True], [True, True]], dtype=jnp.bool_)
+    )
+    workspace64 = _make_binary_task_workspace(indptr64)
+    workspace32 = _make_binary_task_workspace(indptr32)
+
+    got = binary_csrmm_indexed(weights, indices, indptr64, perm, matrix, shape=(2, 3),
+                               transpose=transpose, backend='cuda_raw', workspace=workspace64)
+    expected = binary_csrmm_indexed(weights, indices, indptr32, perm, matrix, shape=(2, 3),
+                                    transpose=transpose, backend='jax_raw', workspace=workspace32)
+
+    assert jnp.allclose(got, expected, rtol=1e-5, atol=1e-5)
