+++ brainevent/_csr/slice_test.py
@@ -22,8 +22,13 @@ import jax.numpy as jnp
 import numpy as np
 import pytest
 
-from brainevent._csr.slice import csr_slice_rows, csr_slice_rows_p
+import brainevent._csr.slice as slice_mod
+from brainevent._csr.slice import csr_slice_rows, csr_slice_rows_grad, csr_slice_rows_p
 from brainevent._csr.main import CSR, CSC
+from brainevent._csr._test_util import (
+    cuda_kwargs, int64_structure, recording_ffi_call, requires_gpu_backend, shape_of,
+)
+from brainevent._test_util import jax_x64_enabled
 
 platform = jax.default_backend()
 SLICE_IMPLEMENTATIONS = tuple(csr_slice_rows_p.available_backends(platform))
@@ -530,3 +535,108 @@ class TestCSCGetitem:
         expected = dense[np.array([1, 5, 9])]
         assert result.shape == (3, n)
         assert jnp.allclose(result, jnp.asarray(expected), atol=1e-5)
+
+
+# ---------------------------------------------------------------------------
+# int64 ``indptr`` policy on the CUDA path.
+#
+# ``indices`` stay int32 (the CUDA ABI is int32-only for coordinates) while
+# ``indptr`` may widen to int64. The generator tests run without a real GPU by
+# stubbing ``load_cuda_file``/``ffi_call``; the ``accepts`` tests need one.
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.parametrize(
+    'factory,args,kwargs',
+    [
+        (
+            slice_mod._csr_slice_rows_cuda_kernel_generator,
+            (),
+            {
+                'outs': [shape_of(jnp.float32, (1, 2))],
+                'data_info': shape_of(jnp.float32),
+                'row_indices_info': shape_of(jnp.int32, (1,)),
+            },
+        ),
+        (
+            slice_mod._csr_slice_rows_grad_cuda_kernel_generator,
+            (),
+            {
+                'outs': [shape_of(jnp.float32)],
+                'ct_info': shape_of(jnp.float32, (1, 2)),
+                'row_indices_info': shape_of(jnp.int32, (1,)),
+            },
+        ),
+    ],
+)
+def test_cuda_kernel_generators_reject_int64_indices_before_loading_cuda(factory, args, kwargs):
+    call_kwargs = cuda_kwargs()
+    call_kwargs.update(kwargs)
+
+    with pytest.raises(TypeError, match="indices with dtype int32"):
+        factory(*args, **call_kwargs)
+
+
+def test_slice_cuda_generators_accept_int64_indptr_without_real_cuda(monkeypatch):
+    ffi_calls = []
+    load_calls = []
+
+    monkeypatch.setattr(slice_mod, "load_cuda_file", lambda path, name: load_calls.append((path, name)))
+    monkeypatch.setattr(slice_mod.jax.ffi, "ffi_call", recording_ffi_call(ffi_calls))
+
+    with jax_x64_enabled():
+        indices = jnp.array([0, 1], dtype=jnp.int32)
+        indptr = jnp.array([0, 2], dtype=jnp.int64)
+
+        slice_kernel = slice_mod._csr_slice_rows_cuda_kernel_generator(
+            **{
+                **cuda_kwargs(indices_dtype=jnp.int32, indptr_dtype=jnp.int64),
+                'outs': [shape_of(jnp.float32, (1, 2))],
+                'data_info': shape_of(jnp.float32, (2,)),
+                'row_indices_info': shape_of(jnp.int32, (1,)),
+            }
+        )
+        slice_kernel(jnp.array([1.0, 2.0]), indices, indptr, jnp.array([0], dtype=jnp.int32))
+
+        slice_grad_kernel = slice_mod._csr_slice_rows_grad_cuda_kernel_generator(
+            **{
+                **cuda_kwargs(indices_dtype=jnp.int32, indptr_dtype=jnp.int64),
+                'outs': [shape_of(jnp.float32, (2,))],
+                'ct_info': shape_of(jnp.float32, (1, 2)),
+                'row_indices_info': shape_of(jnp.int32, (1,)),
+            }
+        )
+        slice_grad_kernel(jnp.array([[1.0, 2.0]]), indices, indptr, jnp.array([0], dtype=jnp.int32))
+
+    assert [name for _, name in load_calls] == ['csr_slice_rows', 'csr_slice_rows']
+    assert [call[0] for call in ffi_calls] == [
+        'csr_slice_rows.csr_slice_rows_fwd_hetero_auto_f32',
+        'csr_slice_rows.csr_slice_rows_grad_auto_f32',
+    ]
+
+
+@requires_gpu_backend
+def test_slice_cuda_accepts_int64_indptr():
+    weights, indices, indptr32 = int64_structure(jnp.int32)
+    indptr64 = indptr32.astype(jnp.int64)
+    rows = jnp.array([1, 0], dtype=jnp.int32)
+
+    got = csr_slice_rows(
+        weights, indices, indptr64, rows, shape=(2, 3), backend='cuda_raw'
+    )
+    expected = jnp.array([[0.0, 3.0, 4.0], [1.0, 0.0, 2.0]], dtype=jnp.float32)
+
+    assert jnp.allclose(got, expected, rtol=1e-5, atol=1e-5)
+
+
+@requires_gpu_backend
+def test_slice_grad_cuda_accepts_int64_indptr():
+    _, indices, indptr32 = int64_structure(jnp.int32)
+    indptr64 = indptr32.astype(jnp.int64)
+    rows = jnp.array([1, 0], dtype=jnp.int32)
+    ct = jnp.array([[10.0, 20.0, 30.0], [1.0, 2.0, 3.0]], dtype=jnp.float32)
+
+    got = csr_slice_rows_grad(ct, indices, indptr64, rows, shape=(2, 3), backend='cuda_raw')
+    expected = jnp.array([1.0, 3.0, 20.0, 30.0], dtype=jnp.float32)
+
+    assert jnp.allclose(got, expected, rtol=1e-5, atol=1e-5)
