--- brainevent/_op/kernix_cpp_test.py
@@ -1,217 +1,0 @@
-# Copyright 2025 BrainX Ecosystem Limited. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-
-"""Kernix C++/CPU FFI end-to-end tests."""
-
-import platform
-
-import jax
-import jax.numpy as jnp
-import numpy as np
-import pytest
-
-import brainevent
-
-if platform.platform().startswith('Windows'):
-    pytest.skip(reason="Windows is not supported yet.", allow_module_level=True)
-
-ADD_ONE_SRC = r"""
-#include "brainevent/common.h"
-
-void add_one_cpu(const BE::Tensor x, BE::Tensor y) {
-    int n = x.numel();
-    const float* in_ptr = static_cast<const float*>(x.data_ptr());
-    float* out_ptr = static_cast<float*>(y.data_ptr());
-    for (int i = 0; i < n; ++i) {
-        out_ptr[i] = in_ptr[i] + 1.0f;
-    }
-}
-"""
-
-SCALE_SRC = r"""
-#include "brainevent/common.h"
-
-void scale_cpu(const BE::Tensor x, BE::Tensor y) {
-    int n = x.numel();
-    const float* in_ptr = static_cast<const float*>(x.data_ptr());
-    float* out_ptr = static_cast<float*>(y.data_ptr());
-    for (int i = 0; i < n; ++i) {
-        out_ptr[i] = in_ptr[i] * 2.0f;
-    }
-}
-"""
-
-MULTI_OUT_SRC = r"""
-#include "brainevent/common.h"
-
-void split_cpu(const BE::Tensor x,
-               BE::Tensor lo, BE::Tensor hi) {
-    int n = x.numel();
-    int half = n / 2;
-    const float* src = static_cast<const float*>(x.data_ptr());
-    float* lo_ptr = static_cast<float*>(lo.data_ptr());
-    float* hi_ptr = static_cast<float*>(hi.data_ptr());
-    for (int i = 0; i < half; ++i) lo_ptr[i] = src[i];
-    for (int i = 0; i < n - half; ++i) hi_ptr[i] = src[half + i];
-}
-"""
-
-
-@pytest.fixture(scope="module")
-def cpu_add_one_module():
-    return brainevent.load_cpp_inline(
-        name="test_cpu_add_one",
-        cpp_sources=ADD_ONE_SRC,
-        functions=["add_one_cpu"],
-        force_rebuild=True,
-    )
-
-
-def test_add_one_cpu_jit(cpu_add_one_module):
-    """CPU kernel works under jax.jit."""
-
-    cpu = jax.devices("cpu")[0]
-
-    @jax.jit
-    def add_one(x):
-        return jax.ffi.ffi_call(
-            "test_cpu_add_one.add_one_cpu",
-            jax.ShapeDtypeStruct(x.shape, x.dtype),
-            vmap_method="broadcast_all",
-        )(x)
-
-    x = jax.device_put(jnp.arange(256, dtype=jnp.float32), cpu)
-    result = add_one(x)
-    expected = np.arange(256, dtype=np.float32) + 1.0
-    np.testing.assert_allclose(np.asarray(result), expected)
-
-
-def test_explicit_dict_form():
-    """CPU kernel with explicit dict-form functions."""
-
-    mod = brainevent.load_cpp_inline(
-        name="test_cpu_scale",
-        cpp_sources=SCALE_SRC,
-        functions={"scale_cpu": ["arg", "ret"]},
-        force_rebuild=True,
-    )
-
-    cpu = jax.devices("cpu")[0]
-    x = jax.device_put(jnp.array([1.0, 2.0, 3.0], dtype=jnp.float32), cpu)
-    result = jax.ffi.ffi_call(
-        "test_cpu_scale.scale_cpu",
-        jax.ShapeDtypeStruct(x.shape, x.dtype),
-        vmap_method="broadcast_all",
-    )(x)
-
-    expected = np.array([2.0, 4.0, 6.0], dtype=np.float32)
-    np.testing.assert_allclose(np.asarray(result), expected)
-
-
-def test_multi_output_cpu():
-    """CPU kernel with two output buffers."""
-    mod = brainevent.load_cpp_inline(
-        name="test_cpu_split",
-        cpp_sources=MULTI_OUT_SRC,
-        functions=["split_cpu"],
-        force_rebuild=True,
-    )
-
-    cpu = jax.devices("cpu")[0]
-    n = 256
-    x = jax.device_put(jnp.arange(n, dtype=jnp.float32), cpu)
-
-    lo, hi = jax.ffi.ffi_call(
-        "test_cpu_split.split_cpu",
-        (
-            jax.ShapeDtypeStruct((n // 2,), jnp.float32),
-            jax.ShapeDtypeStruct((n // 2,), jnp.float32),
-        ),
-        vmap_method="broadcast_all",
-    )(x)
-
-    np.testing.assert_allclose(
-        np.asarray(lo), np.arange(n // 2, dtype=np.float32)
-    )
-    np.testing.assert_allclose(
-        np.asarray(hi), np.arange(n // 2, n, dtype=np.float32)
-    )
-
-
-# ---------------------------------------------------------------------------
-# Host-side check propagation tests
-# ---------------------------------------------------------------------------
-
-# A CPU kernel that fails a host-side invariant.  ``x.numel()`` is always
-# non-negative, so ``x.numel() < 0`` is always false and the check fires at
-# runtime (the compiler cannot prove it away from the source alone).
-CHECK_FAIL_SRC = r"""
-#include "brainevent/common.h"
-
-void check_fail_cpu(const BE::Tensor x, BE::Tensor y) {
-    BE_CHECK(x.numel() < 0) << "intentional failure: numel=" << x.numel();
-    float* out_ptr = static_cast<float*>(y.data_ptr());
-    out_ptr[0] = 1.0f;  // unreachable
-}
-"""
-
-
-def test_host_check_failure_raises_not_aborts():
-    """A failing BE_CHECK surfaces as a Python exception (no SIGABRT)."""
-    mod = brainevent.load_cpp_inline(
-        name="test_cpu_check_fail",
-        cpp_sources=CHECK_FAIL_SRC,
-        functions=["check_fail_cpu"],
-        force_rebuild=True,
-    )
-
-    cpu = jax.devices("cpu")[0]
-    x = jax.device_put(jnp.array([1.0, 2.0, 3.0], dtype=jnp.float32), cpu)
-
-    with pytest.raises(Exception) as excinfo:
-        result = jax.ffi.ffi_call(
-            "test_cpu_check_fail.check_fail_cpu",
-            jax.ShapeDtypeStruct(x.shape, x.dtype),
-            vmap_method="broadcast_all",
-        )(x)
-        jax.block_until_ready(result)
-
-    # The diagnostic from check.h propagates through the FFI error.
-    message = str(excinfo.value)
-    assert "CHECK FAILED" in message or "intentional failure" in message, message
-
-
-def test_process_survives_after_check_failure():
-    """After a check failure, the interpreter is still usable (not aborted)."""
-    mod = brainevent.load_cpp_inline(
-        name="test_cpu_check_fail2",
-        cpp_sources=CHECK_FAIL_SRC.replace("check_fail_cpu", "check_fail_cpu2"),
-        functions=["check_fail_cpu2"],
-        force_rebuild=True,
-    )
-    cpu = jax.devices("cpu")[0]
-    x = jax.device_put(jnp.array([1.0, 2.0, 3.0], dtype=jnp.float32), cpu)
-
-    with pytest.raises(Exception):
-        jax.block_until_ready(
-            jax.ffi.ffi_call(
-                "test_cpu_check_fail2.check_fail_cpu2",
-                jax.ShapeDtypeStruct(x.shape, x.dtype),
-                vmap_method="broadcast_all",
-            )(x)
-        )
-
-    # Still alive: a normal JAX computation completes after the failure.
-    assert float(jnp.sum(jnp.arange(5.0))) == 10.0
